+++ meerkat-dbm/src/file-manager/__tests__/parallel-memory-file-manager.spec.ts
@@ -1,5 +1,5 @@
 import log from 'loglevel';
-import { InstanceManager } from '../../dbm/__test__/mock';
+import { InstanceManager } from '../../dbm/__test__/mock.spec';
 import { FileBufferStore } from '../file-manager-type';
 import { ParallelMemoryFileManager } from '../memory/parallel-memory-file-manager';
